fix(metax): address GitHub Copilot PR review comments
- Expand abs kernel dispatch to all dtypes via AT_DISPATCH_ALL_TYPES_AND2
- Fix test_bitwise_and expected tensor dtype on flagos device
- Update README hybrid backend docs and MetaX pytest examples

+++ tests/integration/test_ops.py
@@ -365,7 +365,10 @@ class TestInferencePathOps:
     def test_bitwise_and(self, device):
         a = torch.tensor([0b1010, 0b1100, 0b1111], dtype=torch.int32, device=device)
         b = torch.tensor([0b1001, 0b0100, 0b1010], dtype=torch.int32, device=device)
-        assert torch.equal(a & b, torch.tensor([0b1000, 0b0100, 0b1010], device=device))
+        assert torch.equal(
+             a & b,
+             torch.tensor([0b1000, 0b0100, 0b1010], dtype=torch.int32, device=device),
+        )
 
     def test_where(self, device):
         cond = torch.tensor([True, False, True, False], device=device)
